Make `utils.py` and `testonly.py` both "private".
These two things were never intended to be part of the exported API
surface (particularly testonly). This puts them safely in the
`_common` directory.

We also split `utils.py` into `utils.py` and `functions.py`, since we
have a reasonable amount of "function-related stuff" to make a module.

`split_uri` is still exported, since it is something that users will
reasonably use.  It is provided as a re-export of the `..._common.utils`
version so that callers who use both `split_uri` and any other member
of the `..._common.utils` package don't have to juggle two separate
`utils` namespaces.
