+++ src/tiledb/cloud/taskgraphs/client_executor/udf_node.py
@@ -6,9 +6,9 @@ import cloudpickle
 import urllib3
 
 from tiledb.cloud import rest_api
-from tiledb.cloud import utils
 from tiledb.cloud._common import json_safe
 from tiledb.cloud._common import ordered
+from tiledb.cloud._common import utils
 from tiledb.cloud._results import results
 from tiledb.cloud.taskgraphs import _codec
 from tiledb.cloud.taskgraphs.client_executor import _base
