+++ tests/taskgraphs/test_taskgraphs.py
@@ -4,7 +4,7 @@ import datetime
 import unittest
 
 import tiledb.cloud.taskgraphs as tg
-from tiledb.cloud import testonly
+from tiledb.cloud._common import testonly
 
 
 class TaskGraphsTest(unittest.TestCase):
