+++ tests/test_basic.py
@@ -10,9 +10,9 @@ import tiledb.cloud
 from tiledb.cloud import array
 from tiledb.cloud import client
 from tiledb.cloud import tasks
-from tiledb.cloud import testonly
 from tiledb.cloud import tiledb_cloud_error
 from tiledb.cloud._common import json_safe
+from tiledb.cloud._common import testonly
 
 
 class BasicTests(unittest.TestCase):
